moved record page
