+++ .github/workflows/build-and-release.yml
@@ -1,0 +1,32 @@
+name: Build APK and Upload to Release
+
+on:
+  push:
+    branches:
+      - master 
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v4
+
+    - name: Set up Flutter
+      uses: subosito/flutter-action@v2
+      with:
+        flutter-version: '3.32.2'  # Change to match your version
+
+    - name: Install dependencies
+      run: flutter pub get
+
+    - name: Build APK
+      run: flutter build apk --release
+
+    - name: Upload APK to GitHub Release
+      uses: ncipollo/release-action@v1
+      with:
+        artifacts: build/app/outputs/flutter-apk/app-release.apk
+        tag: ${{ github.ref_name }}  
+        token: ${{ secrets.GITHUB_TOKEN }}
+        allowUpdates: true
