+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "cmake.sourceDirectory": "/home/oli/git/stride/linux"
+}
