+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'record_page.dart';
+import 'pages/record_page.dart';
 
 void main() {
   runApp(const MyApp());
